+++ src/kqueue.rs
@@ -1,12 +1,13 @@
-//! Bindings to kqueue (macOS, iOS, tvOS, watchOS, FreeBSD, NetBSD, OpenBSD, DragonFly BSD).
+//! Bindings to kqueue (macOS, iOS, tvOS, watchOS, visionOS, FreeBSD, NetBSD, OpenBSD, DragonFly BSD).
 
 use std::collections::HashSet;
 use std::io;
 use std::os::unix::io::{AsFd, AsRawFd, BorrowedFd, OwnedFd, RawFd};
 use std::sync::RwLock;
-use std::time::Duration;
+use std::time::Instant;
 
-use rustix::event::kqueue;
+use rustix::buffer::spare_capacity;
+use rustix::event::{kqueue, Timespec};
 use rustix::io::{fcntl_setfd, Errno, FdFlags};
 
 use crate::{Event, PollMode};
@@ -62,6 +63,7 @@ impl Poller {
         // Register the notification pipe.
         poller.notify.register(&poller)?;
 
+        #[cfg(feature = "tracing")]
         tracing::trace!(
             kqueue_fd = ?poller.kqueue_fd.as_raw_fd(),
             "new"
@@ -93,6 +95,7 @@ impl Poller {
 
     /// Modifies an existing file descriptor.
     pub fn modify(&self, fd: BorrowedFd<'_>, ev: Event, mode: PollMode) -> io::Result<()> {
+        #[cfg(feature = "tracing")]
         let span = if !self.notify.has_fd(fd) {
             let span = tracing::trace_span!(
                 "add",
@@ -104,6 +107,7 @@ impl Poller {
         } else {
             None
         };
+        #[cfg(feature = "tracing")]
         let _enter = span.as_ref().map(|s| s.enter());
 
         self.has_source(SourceId::Fd(fd.as_raw_fd()))?;
@@ -151,7 +155,12 @@ impl Poller {
             let changelist = changelist.as_ref();
 
             unsafe {
-                kqueue::kevent(&self.kqueue_fd, changelist, &mut eventlist, None)?;
+                kqueue::kevent_timespec(
+                    &self.kqueue_fd,
+                    changelist,
+                    spare_capacity(&mut eventlist),
+                    None,
+                )?;
             }
         }
 
@@ -225,23 +234,40 @@ impl Poller {
         self.remove_source(SourceId::Fd(fd.as_raw_fd()))
     }
 
-    /// Waits for I/O events with an optional timeout.
-    pub fn wait(&self, events: &mut Events, timeout: Option<Duration>) -> io::Result<()> {
+    /// Waits for I/O events with an optional deadline.
+    pub fn wait_deadline(&self, events: &mut Events, deadline: Option<Instant>) -> io::Result<()> {
+        #[cfg(feature = "tracing")]
         let span = tracing::trace_span!(
             "wait",
             kqueue_fd = ?self.kqueue_fd.as_raw_fd(),
-            ?timeout,
+            ?deadline,
         );
+        #[cfg(feature = "tracing")]
         let _enter = span.enter();
+
+        let timeout = deadline.map(|deadline| deadline.saturating_duration_since(Instant::now()));
+
+        // Timeout for kevent. In case of overflow, use no timeout.
+        let timeout = match timeout {
+            Some(t) => Timespec::try_from(t).ok(),
+            None => None,
+        };
 
         // Wait for I/O events.
         let changelist = [];
-        let eventlist = &mut events.list;
-        let res = unsafe { kqueue::kevent(&self.kqueue_fd, &changelist, eventlist, timeout)? };
+        let _res = unsafe {
+            kqueue::kevent_timespec(
+                &self.kqueue_fd,
+                &changelist,
+                spare_capacity(&mut events.list),
+                timeout.as_ref(),
+            )?
+        };
 
+        #[cfg(feature = "tracing")]
         tracing::trace!(
             kqueue_fd = ?self.kqueue_fd.as_raw_fd(),
-            ?res,
+            res = ?_res,
             "new events",
         );
 
@@ -253,10 +279,12 @@ impl Poller {
 
     /// Sends a notification to wake up the current or next `wait()` call.
     pub fn notify(&self) -> io::Result<()> {
+        #[cfg(feature = "tracing")]
         let span = tracing::trace_span!(
             "notify",
             kqueue_fd = ?self.kqueue_fd.as_raw_fd(),
         );
+        #[cfg(feature = "tracing")]
         let _enter = span.enter();
 
         self.notify.notify(self).ok();
@@ -278,10 +306,12 @@ impl AsFd for Poller {
 
 impl Drop for Poller {
     fn drop(&mut self) {
+        #[cfg(feature = "tracing")]
         let span = tracing::trace_span!(
             "drop",
             kqueue_fd = ?self.kqueue_fd.as_raw_fd(),
         );
+        #[cfg(feature = "tracing")]
         let _enter = span.enter();
 
         let _ = self.notify.deregister(self);
@@ -397,15 +427,13 @@ pub(crate) fn mode_to_flags(mode: PollMode) -> kqueue::EventFlags {
 #[cfg(any(
     target_os = "freebsd",
     target_os = "dragonfly",
-    target_os = "macos",
-    target_os = "ios",
-    target_os = "tvos",
-    target_os = "watchos",
+    target_vendor = "apple",
 ))]
 mod notify {
     use super::Poller;
     use rustix::event::kqueue;
     use std::io;
+    #[cfg(feature = "tracing")]
     use std::os::unix::io::BorrowedFd;
 
     /// A notification pipe.
@@ -471,6 +499,7 @@ mod notify {
         }
 
         /// Whether this raw file descriptor is associated with this pipe.
+        #[cfg(feature = "tracing")]
         pub(super) fn has_fd(&self, _fd: BorrowedFd<'_>) -> bool {
             false
         }
@@ -480,17 +509,16 @@ mod notify {
 #[cfg(not(any(
     target_os = "freebsd",
     target_os = "dragonfly",
-    target_os = "macos",
-    target_os = "ios",
-    target_os = "tvos",
-    target_os = "watchos",
+    target_vendor = "apple",
 )))]
 mod notify {
     use super::Poller;
     use crate::{Event, PollMode, NOTIFY_KEY};
     use std::io::{self, prelude::*};
+    #[cfg(feature = "tracing")]
+    use std::os::unix::io::BorrowedFd;
     use std::os::unix::{
-        io::{AsFd, AsRawFd, BorrowedFd},
+        io::{AsFd, AsRawFd},
         net::UnixStream,
     };
 
@@ -560,6 +588,7 @@ mod notify {
         }
 
         /// Whether this raw file descriptor is associated with this pipe.
+        #[cfg(feature = "tracing")]
         pub(super) fn has_fd(&self, fd: BorrowedFd<'_>) -> bool {
             self.read_stream.as_raw_fd() == fd.as_raw_fd()
         }